Rename and cleanup structure a bit
