+++ src/RuntimeModule.ts
@@ -19,11 +19,7 @@ import { EventEmitter } from "events";
 import { ModuleApi } from "./ModuleApi";
 import { PlainSubstitution } from "./types/translations";
 
-import { AllExtensions } from "./lifecycles/types";
-import { 
-    IProvideCryptoSetupExtensions,
-    IProvideOtherExtensions
- } from "./lifecycles/SecurityLifecycle";
+import { AllExtensions } from "./ProxiedExtensions";
 
 // TODO: Type the event emitter with AnyLifecycle (extract TypedEventEmitter from js-sdk somehow?)
 // See https://github.com/matrix-org/matrix-react-sdk-module-api/issues/4
@@ -32,8 +28,6 @@ import {
  * Represents a module which is loaded at runtime. Modules which implement this class
  * will be provided information about the application state and can react to it.
  */
-
-
     
 export abstract class RuntimeModule extends EventEmitter {
 
@@ -44,7 +38,7 @@ export abstract class RuntimeModule extends EventEmitter {
 
         this.extensions = {
             cryptoSetup: undefined,
-            other: undefined
+            experimental: undefined
         }
     }
